refactor: user-role.ts 파일 생성(USER_ROLE,type UserRole 이동)
