+++ src/features/auth/index.ts
@@ -1,5 +1,3 @@
 export { OAuthCallback } from './pages/oauth-callback';
 
 export { authApi } from './api/auth-api';
-
-export { type UserRole } from './types/auth-types';
