+++ src/shared/constants/index.ts
@@ -2,4 +2,6 @@ export { createBaseQuery, API_BASE_URL } from './api-config';
 
 export { ROUTES } from './routes';
 
-export { NAV_ITEMS, INFO_LINKS, PARTNER_LINKS, USER_ROLE } from './layout';
+export { NAV_ITEMS, INFO_LINKS, PARTNER_LINKS } from './layout';
+
+export { USER_ROLE, type UserRole } from './user-role';
