+++ src/shared/constants/layout.ts
@@ -26,12 +26,3 @@ export const PARTNER_LINKS = [
     isBold: true,
   },
 ] as const;
-
-// role
-export const USER_ROLE = {
-  TEMP: 'TEMP',
-  CUSTOMER: 'CUSTOMER',
-  ARTIST: 'ARTIST',
-  SHOP: 'SHOP',
-  ADMIN: 'ADMIN',
-} as const;
