+++ src/shared/constants/user-role.ts
@@ -1,0 +1,9 @@
+export const USER_ROLE = {
+  TEMP: 'TEMP',
+  CUSTOMER: 'CUSTOMER',
+  ARTIST: 'ARTIST',
+  SHOP: 'SHOP',
+  ADMIN: 'ADMIN',
+} as const;
+
+export type UserRole = 'TEMP' | 'CUSTOMER' | 'ARTIST' | 'SHOP' | 'ADMIN';
